requirements.txt
APScheduler==3.11.0
python-dotenv==1.2.2
python-telegram-bot==22.7
